@@ -6,7 +6,7 @@ const routes: Routes = [];
 @NgModule({
   imports: [RouterModule.forRoot(routes, {
     anchorScrolling: 'enabled'
-  })],
+})],
   exports: [RouterModule]
 })
 export class AppRoutingModule { }
