@@ -1,10 +1,10 @@
 import {Component, OnInit, ViewEncapsulation} from "@angular/core";
 
 @Component({
-  selector: 'rm-info-item',
-  templateUrl: './info-item.component.html',
-  styleUrls: ['./info-item.component.scss'],
-  encapsulation: ViewEncapsulation.None
+    selector: 'rm-info-item',
+    templateUrl: './info-item.component.html',
+    styleUrls: ['./info-item.component.scss'],
+    encapsulation: ViewEncapsulation.None
 })
 export class InfoItemComponent implements OnInit {
 
