@@ -1,9 +1,12 @@
 import { Component, OnInit } from '@angular/core';
+import { InfoComponent } from '../../components/info/info.component';
+import { InfoItemComponent } from '../../components/info-item/info-item.component';
 
 @Component({
-  selector: 'rm-education',
-  templateUrl: './education.component.html',
-  styleUrls: ['./education.component.scss']
+    selector: 'rm-education',
+    templateUrl: './education.component.html',
+    styleUrls: ['./education.component.scss'],
+    imports: [InfoComponent, InfoItemComponent]
 })
 export class EducationComponent implements OnInit {
     dissertationUrl = `${window.location.origin}/RichardMuscat-320ProjectReport-2012.pdf`;
