@@ -1,18 +1,16 @@
 import { Directive, ElementRef, EventEmitter, Injectable, HostListener, Input, Output } from '@angular/core';
 
-@Directive({
-  selector: '[rmScrollSpy]'
-})
+@Directive({ selector: '[rmScrollSpy]' })
 export class ScrollSpyDirective {
 
   @Output() public sectionChange = new EventEmitter<string>();
-  private currentSection: string;
+  private currentSection!: string;
 
   constructor(public _el: ElementRef) {}
 
   @HostListener('window:scroll', ['$event'])
   onScroll(event: any) {
-    let currentSection: string;
+    let currentSection = '';
     const children = this._el.nativeElement.children;
     const scrollTop = event.target.scrollingElement.scrollTop;
     console.log(scrollTop);
